H27MPJ {fmsb}	R Documentation

Tentative model population of Japan in Heisei 27 (2015)

Description

The data gives the age-class specific model population of Japan in smoothed Heisei 27 (2015) to calculate directly adjusted mortality rate.

Usage

 H27MPJ 

Format

A named vector containing 21 observations, where names show age-classes.

Source

https://www.mhlw.go.jp/content/12601000/000638712.pdf

References

Tamura K. (2008) How do we die?: death date from vital statistics of the Japanese population. The Waseda study of politics and public law, 87: 27-57.

[Package fmsb version 0.7.6 Index]