IndexOfDissimilarity {fmsb}	R Documentation

Index of dissimilarity (ID)

Description

Index of dissimilarity between the 2 age-distributions.

Usage

 IndexOfDissimilarity(X, Y)

Arguments

X
A vector of age-specific standard populations (or percentage) for each age.

Y
A vector of age-specific target populations (or percentage) for each age.

Value

ID
Index of dissimilarity, which is a half of sum of absolute differences of percentages for each age, where NA is automatically treated as 0.

Author(s)

Minato Nakazawa minatonakazawa@gmail.com https://minato.sip21c.org/

References

Preston SH, Heuveline P, Guillot M (2001) Demography: Measuring and Modeling Population Processes. Blackwell Publishing, Oxford.

Newell C (1988) Methods and Models in Demography. The Guilford Press, New York.

Rowland DT (2003) Demographic methods and concepts. Oxford Univ. Press, Oxford.

Examples

 # Dissimilarities of Japanese population structure were increasing
 # from 1960-1980 (0.132) to 1980-2000 (0.156).
 IndexOfDissimilarity(Jpopl$M1980+Jpopl$F1980, Jpopl$M2000+Jpopl$F2000)
 IndexOfDissimilarity(Jpopl$M1980+Jpopl$F1980, Jpopl$M1960+Jpopl$F1960)

[Package fmsb version 0.7.6 Index]