RDMH {fmsb}	R Documentation

Calculate pooled risk difference and its confidence intervals with Mantel-Haenszel's method

Description

Calculate pooled risk difference and its confidence intervals with Mantel-Haenszel's method.

Usage

 RDMH(XTAB, conf.level=0.9) 

Arguments

XTAB
A matrix with 4 columns. The first column is the number of disease occurrence in exposed cohort. The second column is the number of disease occurrence in unexposed cohort. The third column is the total number of exposed cohort. The forth column is the total number of unexposed cohort. Rows should be composed of different strata or studies.

conf.level
Probability for confidence intervals. Default is 0.9.

Value

estimate
Calculated point estimate of pooled risk difference with Manterl-Haenszel's method.

conf.int
A numeric vector of length 2 to give upper/lower limit of confidence intervals.

conf.level
Simply return the value of given conf.level.

Author(s)

Minato Nakazawa minatonakazawa@gmail.com https://minato.sip21c.org/

References

Rothman KJ (2012) Epidemiology: An Introduction. 2nd Ed., Oxford University Press, Oxford.

Examples

# Table 10-3 of Rothman's textbook (Chapter 10).
RDMH(matrix(c(8, 5, 106, 120, 22, 16, 98, 85), 2, 4, byrow=TRUE), conf.level=0.9)

[Package fmsb version 0.7.6 Index]