gstem {fmsb}	R Documentation

Draw the stem and leaf plot in a graphic device.

Description

Capture the output of stem() function and plot them into graphic devices. However, the result of setting scale parameter as 2 may be controversial.

Usage

 gstem(X, scale) 

Arguments

X
A numeric vector.

scale
Parameter to control plot length of graph. Default is 1.

Author(s)

Minato Nakazawa minatonakazawa@gmail.com https://minato.sip21c.org/

Examples

 x <- rnorm(100, 10, 1)
 stem(x)
 stem(x, 2)
 layout(t(1:2))
 gstem(x)
 gstem(x, 2)

[Package fmsb version 0.7.6 Index]