mhchart {fmsb}	R Documentation

Draw maternity history charts.

Description

To compare the maternity histories among several human populations, this kind of graph is useful, inspired by Wood JW (1994) "Dynamics of Human Reproduction", Aldine de Gruyter, New York.

Usage

mhchart(LIST, XLIM=c(15,45), COL="black", FILL="white", BWD=1, ...)

Arguments

LIST
The list of groups with their maternity histories from first birth to the last birth. Usually the first childbirth age is estimated as median by Kaplan-Meier method, the second childbirth age was given by adding the median of first birth intervals to the first childbirth age by Kaplan-Meier method, and so on.

XLIM
The limit of x axis, which means the range of reproductive ages. Default is 15 and 45.

COL
The border color. Default is black.

FILL
The painting color. Default is white.

BWD
The line width of the boxes. Default is 1.

...
Other parameters handed to barplot() to draw axes and background.

Value

No value is returned.

Author(s)

Minato Nakazawa minatonakazawa@gmail.com https://minato.sip21c.org/

Examples

 Developing <- c(18, 21, 24, 27, 30, 33.5, 37)
 Hutterite <- c(23, 25, 27, 29, 31, 33, 35, 37, 39)
 Gainj <- c(27, 31, 35, 39)
 Japan <- c(29, 34)
 x <- list(
  Developing=Developing,
  Hutterite=Hutterite,
  Gainj=Gainj,
  Japan=Japan)
 mhchart(rev(x), COL="blue", FILL="pink", BWD=2, XLIM=c(15, 45),
  main="Maternity histories for selected populations",
  xlab="Maternal age (years)")

[Package fmsb version 0.7.6 Index]